Selectable text-line density, font-specific menu highlight centering
1.
There is a new "Setup/Set Line Spacing" menu option.
This allows selection between...

 DENSE = 14 text lines in menu
 MEDIUM = 12 text lines in menu
 SPARSE = 10 text lines in menu

The menu system adjusts accordingly.

The MEDIUM is actually quite nice, and perhaps should be the default in future.

2.
The highlight line for current menu item is now adusted slightly on a per-font basis.
This allows the inconsistencies in vertical positioning of the fonts to be catered
for and the menu line is now much better centered for the SMALL CAPS font.

KNOWN BUGS

1. When the main menu is on 2 pages, the text/line colour of the 2nd page item is wrong
2. Ditto, 2nd page... when choosing Setup we first have to go through two "blank" menus