More work on the line spacing and communication -> 2600
* new ZP var 'lineCounter' counts lines in the kernel for use in menu navigation

The easiest way seemed to be to just manage the variable in the kernel, such that
at the end of every frame it has the correct number of lines on the menu/screen.

This value is used when you press UP in menu navigation. If at the top already
(line 0) then UP takes you to the previous page AND the last line on that page.
COuld not use the #elements on page from online value, as that is incorrect for
the last page which is truncated.  Although a bit of a 'hack' in that it's not
as elegant as it could be, this solution  works well enough.

The idea of UP going to the last item on the previous page is consistency; if
you are going down by one line when pressing DOWN (and you ALWAYS go down by
one line, even if the line happens to be on the next page), then UP should be
consistent, and it is now - you ALWAYS go up by one line, even if it is on the
previous page.

Menus behave slightly differently to viewing file contents, though - UP will
always go to the first item on the previous menu page.  It could be made more
consistent, but seems to work pretty well as-is.

KNOWN ISSUES:

Now that there are more menu items in the main menu, and the # lines can be
adjusted to smaller, there are often menu items on page #2.  There are a couple
of issues known...

1) colour of menu items on page 2 may be incorrect. I thought I fixed this, but
it seems to have returned.

2) Sometimes when selecting an item on page 2, you have to go through a few
blank menus before you see the menu you want (see below).

3) I saw an item disappear between pages (I think this is the server having an
incorrect page size concept and feeding incorrect lines.

Note:

I am currently starting a complete rewrite of the menu handler. As previously
discussed it is ugly and hard to follow. I want to get rid of those bad "blank"
menu pages like for example the exiting from search, and the ones that
now seem to appear (occasionally?) on the selection of page 2 menu items.

CONSEQUENTLY THIS CODE IS NOT READY FOR RELEASE (EVEN TO DEBUG TESTERS) AND
IS A WORK-IN-PROGRESS!