Rewrites to save ROM/RAM space
I've changed a few bits of code to be more efficient and save ROM.
The character-array 'status_message' is now properly defined to not pad
with extra characters. It no longer lives in .flash0 as I can't see how
an array of this type can be placed there.

In any case, ROM savings almost 1K.  Most of the action is in the TV_Mode,
Line_spacing and Font_style functions/code.